# internal/docs/generated/swagger.yaml
basePath: /api
definitions:
  apicommon.APIErrorBody:
    properties:
      code:
        description: 错误码
        type: string
      details:
        type: object
      message:
        description: 错误描述
        type: string
    type: object
  apicommon.APIErrorResponse:
    properties:
      error:
        allOf:
        - $ref: '#/definitions/apicommon.APIErrorBody'
        description: 错误体
    type: object
...
  v2.V2EfficiencyAPIResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/v2.V2EfficiencyResponse'
        description: 响应数据
      meta:
        allOf:
        - $ref: '#/definitions/v2.V2Meta'
        description: 响应元信息
    type: object
  v2.V2EfficiencyBuckets:
    properties:
      speed_10kmh:
        items:
          $ref: '#/definitions/v2.V2EfficiencySpeedBucket'
        type: array
      temperature_5c:
        items:
          $ref: '#/definitions/v2.V2EfficiencyTempBucket'
        type: array
    type: object
  v2.V2EfficiencyResponse:
    properties:
      buckets:
        $ref: '#/definitions/v2.V2EfficiencyBuckets'
      summary:
        $ref: '#/definitions/v2.V2EfficiencySummary'
    type: object
  v2.V2EfficiencySpeedBucket:
    properties:
      bucket:
        type: number
      consumption:
        type: number
      distance:
        description: 距离 (km)
        type: number
      drive_count:
        description: 行程数
        type: integer
    type: object
  v2.V2EfficiencySummary:
    properties:
      avg_outside_temp:
        type: number
      consumption_overhead:
        description: 能耗开销 (Wh/km)
        type: number
      drive_distance:
        type: number
      drive_duration:
        type: number
      energy_consumed_drives:
        type: number
      energy_consumed_idle:
        type: number
      gross_consumption:
        description: 毛能耗 (Wh/km)
        type: number
      net_consumption:
        type: number
    type: object
  v2.V2EfficiencyTempBucket:
    properties:
      avg_speed:
        description: 平均速度 (km/h)
        type: number
      bucket:
        type: number
      consumption:
        type: number
      distance:
        description: 距离 (km)
        type: number
      drive_count:
        description: 行程数
        type: integer
    type: object
...
  v2.V2Meta:
    properties:
      applied_filters:
        additionalProperties:
          type: string
        description: |-
          AppliedFilters echoes back the query-string filters the server actually
          honored (audit §5.2 / spec §5.2). Helps clients debug "why did I get this
          payload?" — they can compare what they sent vs. what the server applied
          (defaults, normalized casing, ignored unknown keys, etc.). Always
          non-nil when populated; serialized as a JSON object.
        type: object
      car_id:
        description: 车辆 ID
        example: 1
        type: integer
      compare:
        description: 对比模式
        enum:
        - none
        - previous_period
        type: string
      end:
        description: 结束时间
        format: date-time
        type: string
      generated_at:
        format: date-time
        type: string
      period:
        description: 聚合周期
        enum:
        - day
        - week
        - month
        - quarter
        - year
        - custom
        - lifetime
        type: string
      start:
        description: 起始时间
        format: date-time
        type: string
      timezone:
        description: 时区
        example: Asia/Shanghai
        type: string
      unit:
        allOf:
        - $ref: '#/definitions/v2.V2Unit'
        description: 单位
    type: object
...
  v2.V2Unit:
    properties:
      consumption:
        example: Wh/km
        type: string
      currency:
        example: CNY
        type: string
      distance:
        description: 距离 (km)
        example: km
        type: string
      duration:
        description: 时长 (秒)
        example: seconds
        type: string
      elevation:
        example: m
        type: string
      energy:
        example: kWh
        type: string
      power:
        example: kW
        type: string
      speed:
        example: km/h
        type: string
      temperature:
        example: C
        type: string
    type: object
...
info:
  contact: {}
  description: TeslaMate 数据查询与分析 API。提供 V1（原始数据）和 V2（聚合分析）两套接口。
  title: TeslaMateApi
  version: "2.0"
paths:
...
  /v2/cars/{CarID}/analytics/efficiency:
    get:
      description: 返回该车的净能耗、毛能耗、能耗开销，以及可选的温度/速度分桶。group_by 仅在 include=buckets 时生效，否则返回
        400 PARAM_DEPENDENCY_VIOLATION。
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 聚合周期
        enum:
        - day
        - week
        - month
        - quarter
        - year
        - custom
        in: query
        name: period
        type: string
      - description: 起始时间（RFC3339）
        in: query
        name: start
        type: string
      - description: 结束时间（RFC3339）
        in: query
        name: end
        type: string
      - description: IANA 时区
        in: query
        name: timezone
        type: string
      - description: 逗号分隔的扩展项：buckets
        example: '"buckets"'
        in: query
        name: include
        type: string
      - description: 分桶维度（仅在 include=buckets 时生效）
        enum:
        - temperature_5c
        - speed_10kmh
        in: query
        name: group_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.V2EfficiencyAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apicommon.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apicommon.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apicommon.APIErrorResponse'
      summary: V2 能效分析
      tags:
      - v2
...
swagger: "2.0"
...